+++ server/dashboard.js
@@ -466,6 +466,7 @@ const validateUpdateClient = [
     'isSendingAlerts',
     'isSendingVitals',
     'language',
+    'status',
   ])
     .trim()
     .notEmpty(),
@@ -519,6 +520,7 @@ async function submitUpdateClient(req, res) {
         data.isSendingAlerts,
         data.isSendingVitals,
         data.language,
+        data.status,
         req.params.id,
       )